src/backend/requirements.txt
# /requirements.txt

flask
python-dotenv
python-jose
flask-cors
six
requests